- Bumps autonomi dependency to 0.5.0 and fixes related breaking changes.

+++ src/service/pointer_service.rs
@@ -14,12 +14,12 @@ pub struct Pointer {
     name: Option<String>,
     content: String,
     address: Option<String>,
-    counter: Option<u32>,
+    counter: Option<u64>,
     cost: Option<String>,
 }
 
 impl Pointer {
-    pub fn new(name: Option<String>, content: String, address: Option<String>, counter: Option<u32>, cost: Option<String>) -> Self {
+    pub fn new(name: Option<String>, content: String, address: Option<String>, counter: Option<u64>, cost: Option<String>) -> Self {
         Pointer { name, content, address, counter, cost } 
     }
 }
